stashing before refactoring param panel
